Fix: 미해결 질문 리스트 반환 API 수정 #351

+++ src/main/java/com/selfrunner/gwalit/domain/board/dto/response/BoardMetaRes.java
@@ -34,7 +34,7 @@ public class BoardMetaRes {
 
     private final QuestionStatus status;
 
-    private final Integer replyCount;
+    private final Long replyCount;
 
     private final LocalDateTime createdAt;
